[Silver III] Title: 1, 2, 3 더하기, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/9095. 1， 2， 3 더하기/1， 2， 3 더하기.cc
@@ -7,7 +7,7 @@ int main(){
     cin.tie(0);
     int T,N;
     cin>>T;
-    long long num[11]={0,};
+    int num[11]={0,};
     num[1]=1;
     num[2]=2;
     num[3]=4;
